Backend/requirements.txt
Django==4.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.6.0
pymysql==1.1.0
cryptography==42.0.5
django-filter==24.2
